PolygonIOCurrencySnapshotGrams
