+++ src/services/fetchers/MetalsDevApiFetcher.ts
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import {inject, injectable} from 'inversify';
 import {Logger} from 'winston';
 
@@ -10,9 +9,9 @@ import {
   FetchedValueType,
 } from '../../types/fetchers.js';
 import {PriceDataRepository} from '../../repositories/PriceDataRepository.js';
-import Settings from '../../types/Settings.js';
 import TimeService from '../TimeService.js';
 import {MetalsDevApiDataRepository} from '../../repositories/fetchers/MetalsDevApiDataRepository.js';
+import {MappingRepository} from '../../repositories/MappingRepository.js';
 
 export interface MetalsDevApiPriceInputParams {
   metal: string;
@@ -21,24 +20,18 @@ export interface MetalsDevApiPriceInputParams {
 
 @injectable()
 export class MetalsDevApiFetcher implements FeedFetcherInterface {
+  @inject(MappingRepository) private mappingRepository!: MappingRepository;
   @inject(MetalsDevApiDataRepository) private metalsDevApiDataRepository!: MetalsDevApiDataRepository;
   @inject(PriceDataRepository) private priceDataRepository!: PriceDataRepository;
   @inject(TimeService) private timeService!: TimeService;
   @inject('Logger') private logger!: Logger;
 
-  private apiKey: string;
-  private timeout: number;
   private logPrefix = `[${FetcherName.MetalsDevApi}]`;
   static fetcherSource = '';
 
-  constructor(@inject('Settings') settings: Settings) {
-    this.apiKey = settings.api.metalsDevApi.apiKey;
-    this.timeout = settings.api.metalsDevApi.timeout;
-  }
-
   async apply(params: MetalsDevApiPriceInputParams[], options: FeedFetcherOptions): Promise<FetcherResult> {
     try {
-      await this.fetchPrices(params);
+      await this.cacheInput(params);
     } catch (e) {
       this.logger.error(`${this.logPrefix} failed: ${(e as Error).message}`);
     }
@@ -60,47 +53,18 @@ export class MetalsDevApiFetcher implements FeedFetcherInterface {
     return result;
   }
 
-  private async fetchPrices(params: MetalsDevApiPriceInputParams[]): Promise<void> {
-    if (params.length != 1) throw new Error(`${this.logPrefix} not a multifetcher: ${params}`);
+  private async cacheInput(params: MetalsDevApiPriceInputParams[]): Promise<void> {
+    const timestamp = this.timeService.apply();
+    const key = `${FetcherName.MetalsDevApi}_cachedParams`;
 
-    const {metal, currency} = params[0];
+    // const cache = await this.mappingRepository.get(key);
+    // const cachedParams = JSON.parse(cache || '{}');
+    const cachedParams: Record<string, number> = {};
 
-    this.logger.debug(`${this.logPrefix} call for: ${metal}/${currency}`);
+    params.forEach((input) => {
+      cachedParams[`${input.metal};${input.currency}`.toLowerCase()] = timestamp;
+    });
 
-    const apiUrl = 'https://api.metals.dev/v1/latest';
-    const url = `${apiUrl}?api_key=${this.apiKey}&currency=${currency}&unit=g`;
-
-    try {
-      const response = await axios.get(url, {
-        timeout: this.timeout,
-        timeoutErrorMessage: `${this.logPrefix} Timeout exceeded: ${url}`,
-      });
-
-      if (response.status !== 200) {
-        this.logger.error(
-          `${this.logPrefix} Error for ${metal}/${currency}: ${response.statusText}. Error: ${response.data}`,
-        );
-        return;
-      }
-
-      const pricePerGram = response.data.metals[metal.toLowerCase()];
-
-      if (!pricePerGram) {
-        this.logger.error(`${this.logPrefix} Missing price for ${metal} in ${currency}`);
-        return;
-      }
-
-      this.logger.debug(`${this.logPrefix} resolved price per gram: ${metal}/${currency}: ${pricePerGram}`);
-
-      await this.metalsDevApiDataRepository.save([
-        {
-          value: pricePerGram,
-          timestamp: this.timeService.apply(),
-          params: params[0],
-        },
-      ]);
-    } catch (error) {
-      this.logger.error(`${this.logPrefix} An error occurred while fetching metal prices: ${error}`);
-    }
+    await this.mappingRepository.set(key, JSON.stringify(cachedParams));
   }
 }
