+++ src/services/fetchers/PolygonIOCurrencySnapshotGramsFetcher.ts
@@ -1,4 +1,5 @@
 import {inject, injectable} from 'inversify';
+import {Logger} from 'winston';
 
 import {
   FetcherName,
@@ -8,15 +9,13 @@ import {
   FetchedValueType,
 } from '../../types/fetchers.js';
 import {PriceDataRepository} from '../../repositories/PriceDataRepository.js';
-import Settings from '../../types/Settings.js';
 import {PolygonIOCurrencySnapshotGramsDataRepository} from '../../repositories/fetchers/PolygonIOCurrencySnapshotGramsDataRepository.js';
-import {BasePolygonIOSnapshotFetcher, SnapshotResponse} from './common/BasePolygonIOSnapshotFetcher.js';
+import {MappingRepository} from '../../repositories/MappingRepository.js';
+import TimeService from '../TimeService.js';
 
 export interface PolygonIOCurrencySnapshotGramsInputParams {
   ticker: string;
 }
-
-type ParsedResponse = {ticker: string; price: number; timestamp: number};
 
 /*
     - fetcher:
@@ -25,31 +24,23 @@ type ParsedResponse = {ticker: string; price: number; timestamp: number};
           ticker: C:XAUUSD
  */
 @injectable()
-export class PolygonIOCurrencySnapshotGramsFetcher
-  extends BasePolygonIOSnapshotFetcher
-  implements FeedFetcherInterface
-{
+export class PolygonIOCurrencySnapshotGramsFetcher implements FeedFetcherInterface {
+  @inject('Logger') protected logger!: Logger;
+  @inject(MappingRepository) private mappingRepository!: MappingRepository;
+  @inject(TimeService) private timeService!: TimeService;
   @inject(PolygonIOCurrencySnapshotGramsDataRepository)
   private pIOCurrencySnapshotGramsDataRepository!: PolygonIOCurrencySnapshotGramsDataRepository;
   @inject(PriceDataRepository) private priceDataRepository!: PriceDataRepository;
 
   static fetcherSource = '';
-
-  constructor(@inject('Settings') settings: Settings) {
-    super();
-    this.apiKey = settings.api.polygonIO.apiKey;
-    this.timeout = settings.api.polygonIO.timeout;
-    this.valuePath = '$.ticker.lastQuote.a';
-
-    this.logPrefix = `[${FetcherName.PolygonIOCurrencySnapshotGrams}]`;
-  }
+  private logPrefix = `[${FetcherName.PolygonIOCurrencySnapshotGrams}]`;
 
   async apply(
     params: PolygonIOCurrencySnapshotGramsInputParams[],
     options: FeedFetcherOptions,
   ): Promise<FetcherResult> {
     try {
-      await this.fetchPrices(params);
+      await this.cacheInput(params);
     } catch (e) {
       this.logger.error(`${this.logPrefix} failed: ${(e as Error).message}`);
     }
@@ -69,61 +60,17 @@ export class PolygonIOCurrencySnapshotGramsFetcher
     return {prices};
   }
 
-  private async fetchPrices(params: PolygonIOCurrencySnapshotGramsInputParams[]): Promise<void> {
-    if (params.length != 1) throw new Error(`${this.logPrefix} not a multifetcher: ${params}`);
+  private async cacheInput(params: PolygonIOCurrencySnapshotGramsInputParams[]): Promise<void> {
+    const timestamp = this.timeService.apply();
+    const key = `${FetcherName.PolygonIOCurrencySnapshotGrams}_cachedParams`;
 
-    const {ticker} = params[0];
+    const cache = await this.mappingRepository.get(key);
+    const cachedParams = JSON.parse(cache || '{}');
 
-    const baseUrl = 'https://api.polygon.io/v2/snapshot/locale/global/markets/forex/tickers';
-    // const url = `${baseUrl}/${ticker}?apiKey=${this.apiKey}`;
-    const url = `${baseUrl}?apiKey=${this.apiKey}`;
-
-    this.logger.debug(`${this.logPrefix} call for ${ticker}`);
-
-    const response = <SnapshotResponse>await this.fetch(url, true);
-    const parsed = this.parseResponse(response, params[0]);
-
-    if (!parsed) return;
-
-    await this.savePrices(parsed);
-  }
-
-  private parseResponse(
-    response: SnapshotResponse,
-    params: PolygonIOCurrencySnapshotGramsInputParams,
-  ): ParsedResponse[] {
-    if (response.error) {
-      this.logger.error(`${this.logPrefix} [${response.status}]: ${response.error} for ${params.ticker}`);
-      return [];
-    }
-
-    return response.tickers
-      .map((ticker, ix) => {
-        const oneOzInGrams = 31.1034; // grams
-        const price = (ticker.lastQuote.a as number) / oneOzInGrams;
-
-        if (isNaN(price)) {
-          this.logger.error(`${this.logPrefix}#${ix} price fail for ${params.ticker}. Computed value gave NaN.`);
-          return;
-        }
-
-        this.logger.debug(`${this.logPrefix}#${ix} price for ${params.ticker} (${ticker.ticker}): ${price}`);
-        return {price, ticker: ticker.ticker, timestamp: Math.trunc(ticker.lastQuote.t / 1e3)};
-      })
-      .filter((d) => d !== undefined) as ParsedResponse[];
-  }
-
-  private async savePrices(parsed: ParsedResponse[]): Promise<void> {
-    const allData = parsed.map((data) => {
-      return {
-        timestamp: data.timestamp,
-        value: data.price,
-        params: {
-          ticker: data.ticker,
-        },
-      };
+    params.forEach((input) => {
+      cachedParams[input.ticker.toLowerCase()] = timestamp;
     });
 
-    await this.pIOCurrencySnapshotGramsDataRepository.save(allData);
+    await this.mappingRepository.set(key, JSON.stringify(cachedParams));
   }
 }
