+++ src/workers/fetchers/PolygonIOCurrencySnapshotGramsService.ts
@@ -1,0 +1,111 @@
+import {inject, injectable} from 'inversify';
+
+import {FetcherName, ServiceInterface} from '../../types/fetchers.js';
+import Settings from '../../types/Settings.js';
+import {PolygonIOCurrencySnapshotGramsDataRepository} from '../../repositories/fetchers/PolygonIOCurrencySnapshotGramsDataRepository.js';
+import {MappingRepository} from '../../repositories/MappingRepository.js';
+import {
+  BasePolygonIOSnapshotFetcher,
+  SnapshotResponse,
+} from '../../services/fetchers/common/BasePolygonIOSnapshotFetcher.js';
+
+export interface PolygonIOCurrencySnapshotGramsInputParams {
+  ticker: string;
+}
+
+type ParsedResponse = {ticker: string; price: number; timestamp: number};
+
+@injectable()
+export class PolygonIOCurrencySnapshotGramsService extends BasePolygonIOSnapshotFetcher implements ServiceInterface {
+  @inject(MappingRepository) private mappingRepository!: MappingRepository;
+  @inject(PolygonIOCurrencySnapshotGramsDataRepository)
+  private pIOCurrencySnapshotGramsDataRepository!: PolygonIOCurrencySnapshotGramsDataRepository;
+
+  static fetcherSource = '';
+
+  constructor(@inject('Settings') settings: Settings) {
+    super();
+    this.apiKey = settings.api.polygonIO.apiKey;
+    this.timeout = settings.api.polygonIO.timeout;
+    this.valuePath = '$.ticker.lastQuote.a';
+
+    this.logPrefix = '[PolygonIOCurrencySnapshotGramsService]';
+  }
+
+  async apply(): Promise<void> {
+    try {
+      await this.fetchPrices(await this.getInput());
+    } catch (e) {
+      this.logger.error(`${this.logPrefix} failed: ${(e as Error).message}`);
+    }
+  }
+
+  private async fetchPrices(params: PolygonIOCurrencySnapshotGramsInputParams[]): Promise<void> {
+    if (params.length != 1) throw new Error(`${this.logPrefix} not a multifetcher: ${params}`);
+
+    const {ticker} = params[0];
+
+    const baseUrl = 'https://api.polygon.io/v2/snapshot/locale/global/markets/forex/tickers';
+    // const url = `${baseUrl}/${ticker}?apiKey=${this.apiKey}`;
+    const url = `${baseUrl}?apiKey=${this.apiKey}`;
+
+    this.logger.debug(`${this.logPrefix} call for ${ticker}`);
+
+    const response = <SnapshotResponse>await this.fetch(url, true);
+    const parsed = this.parseResponse(response, params[0]);
+
+    if (!parsed) return;
+
+    await this.savePrices(parsed);
+  }
+
+  private parseResponse(
+    response: SnapshotResponse,
+    params: PolygonIOCurrencySnapshotGramsInputParams,
+  ): ParsedResponse[] {
+    if (response.error) {
+      this.logger.error(`${this.logPrefix} [${response.status}]: ${response.error} for ${params.ticker}`);
+      return [];
+    }
+
+    return response.tickers
+      .map((ticker, ix) => {
+        const oneOzInGrams = 31.1034; // grams
+        const price = (ticker.lastQuote.a as number) / oneOzInGrams;
+
+        if (isNaN(price)) {
+          this.logger.error(`${this.logPrefix}#${ix} price fail for ${params.ticker}. Computed value gave NaN.`);
+          return;
+        }
+
+        this.logger.debug(`${this.logPrefix}#${ix} price for ${params.ticker} (${ticker.ticker}): ${price}`);
+        return {price, ticker: ticker.ticker, timestamp: Math.trunc(ticker.lastQuote.t / 1e3)};
+      })
+      .filter((d) => d !== undefined) as ParsedResponse[];
+  }
+
+  private async savePrices(parsed: ParsedResponse[]): Promise<void> {
+    const allData = parsed.map((data) => {
+      return {
+        timestamp: data.timestamp,
+        value: data.price,
+        params: {
+          ticker: data.ticker,
+        },
+      };
+    });
+
+    await this.pIOCurrencySnapshotGramsDataRepository.save(allData);
+  }
+
+  private async getInput(): Promise<PolygonIOCurrencySnapshotGramsInputParams[]> {
+    const key = `${FetcherName.PolygonIOCurrencySnapshotGrams}_cachedParams`;
+
+    const cache = await this.mappingRepository.get(key);
+    const cachedParams = JSON.parse(cache || '{}');
+
+    return Object.keys(cachedParams).map((ticker) => {
+      return {ticker};
+    });
+  }
+}
